refactor: standardize poster signup screen path and update tests
