+++ lib/features/screens/poster_signup_screen/steps/step_2_personalize.dart
